Move methods over to fragment
